requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
pynput==1.7.3
pyobjc-core==7.3; python_version >= '3.6'
pyobjc-framework-cocoa==7.3; python_version >= '3.6'
pyobjc-framework-quartz==7.3; sys_platform == 'darwin'
python-dotenv==0.18.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sockets==1.0.0
